app_local/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
loguru==0.7.2
ffmpeg-python==0.2.0
requests==2.32.3
pydantic==2.9.1
